# sliceop/f64s: improve performances of Take