Reduce memory allocation:
- from populating the map and
- from using sort.IntsAreSorted.

also removed a loop over indices by removing sort.IntsAreSorted.

```
name                old time/op    new time/op    delta
Take/Len=2-4          63.5ns ± 3%     6.1ns ± 1%   -90.45%  (p=0.000 n=18+18)
Take/Len=4-4          82.7ns ± 2%     6.9ns ± 2%   -91.64%  (p=0.000 n=19+20)
Take/Len=8-4           116ns ± 2%       8ns ± 2%   -93.33%  (p=0.000 n=18+20)
Take/Len=128-4        4.03µs ± 1%    0.07µs ± 2%   -98.24%  (p=0.000 n=20+19)
Take/Len=1024-4       31.8µs ± 1%     0.5µs ± 3%   -98.44%  (p=0.000 n=17+19)
Take/Len=1048576-4    57.2ms ± 2%     3.1ms ± 4%   -94.61%  (p=0.000 n=19+20)

name                old alloc/op   new alloc/op   delta
Take/Len=2-4           32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)
Take/Len=4-4           32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)
Take/Len=8-4           32.0B ± 0%      0.0B       -100.00%  (p=0.000 n=20+20)
Take/Len=128-4        1.49kB ± 0%    0.00kB       -100.00%  (p=0.000 n=20+20)
Take/Len=1024-4       11.0kB ± 0%     0.0kB       -100.00%  (p=0.000 n=20+20)
Take/Len=1048576-4    11.2MB ± 0%     0.0MB       -100.00%  (p=0.000 n=18+20)

name                old allocs/op  new allocs/op  delta
Take/Len=2-4            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
Take/Len=4-4            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
Take/Len=8-4            1.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
Take/Len=128-4          3.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
Take/Len=1024-4         6.00 ± 0%      0.00       -100.00%  (p=0.000 n=20+20)
Take/Len=1048576-4      19.0 ± 0%       0.0       -100.00%  (p=0.000 n=20+20)
```