@rmadar 2020-04-24
great! the way to handle the duplicates and ordering is nice (and indeed removing two loops is quite cool!). I guess the tests are ok, but I have the impression you don't fill `src[index[0]]`, and then you miss the first element, no?