@sbinet 2020-04-24
FYI, testing with this change:
```diff
diff --git a/sliceop/f64s/f64s.go b/sliceop/f64s/f64s.go
index 347e2f36..b45a9961 100644
--- a/sliceop/f64s/f64s.go
+++ b/sliceop/f64s/f64s.go
@@ -97,18 +97,19 @@ func Take(dst, src []float64, indices []int) []float64 {
                return dst
        }
 
-       dst[0] = src[indices[0]]
-       for i := 1; i < len(indices); i++ {
-               v0 := indices[i-1]
-               v1 := indices[i]
+       n := len(indices) - 1
+       for i := 0; i < n; i++ {
+               v0 := indices[i]
+               v1 := indices[i+1]
                switch {
                case v0 == v1:
                        panic(errDuplicateIndices)
                case v0 > v1:
                        panic(errSortedIndices)
                }
-               dst[i] = src[v1]
+               dst[i] = src[v0]
        }
+       dst[n] = src[indices[n]]
 
        return dst
 }
```
(compared to the proposed change in 7e55fec)

lead to this performance improvement:
```
$> benchstat new4.log new7.log
name                old time/op    new time/op    delta
Take/Len=2-4          6.06ns ± 1%    5.86ns ± 1%   -3.34%  (p=0.000 n=18+20)
Take/Len=4-4          6.92ns ± 2%    6.83ns ± 1%   -1.27%  (p=0.000 n=20+17)
Take/Len=8-4          7.76ns ± 2%    7.56ns ± 1%   -2.60%  (p=0.000 n=20+20)
Take/Len=128-4        71.1ns ± 2%    82.9ns ± 1%  +16.59%  (p=0.000 n=19+20)
Take/Len=1024-4        495ns ± 3%     587ns ± 2%  +18.74%  (p=0.000 n=19+19)
Take/Len=1048576-4    3.08ms ± 4%    1.95ms ± 2%  -36.76%  (p=0.000 n=20+18)

name                old alloc/op   new alloc/op   delta
Take/Len=2-4           0.00B          0.00B          ~     (all equal)
Take/Len=4-4           0.00B          0.00B          ~     (all equal)
Take/Len=8-4           0.00B          0.00B          ~     (all equal)
Take/Len=128-4         0.00B          0.00B          ~     (all equal)
Take/Len=1024-4        0.00B          0.00B          ~     (all equal)
Take/Len=1048576-4     0.00B          0.00B          ~     (all equal)

name                old allocs/op  new allocs/op  delta
Take/Len=2-4            0.00           0.00          ~     (all equal)
Take/Len=4-4            0.00           0.00          ~     (all equal)
Take/Len=8-4            0.00           0.00          ~     (all equal)
Take/Len=128-4          0.00           0.00          ~     (all equal)
Take/Len=1024-4         0.00           0.00          ~     (all equal)
Take/Len=1048576-4      0.00           0.00          ~     (all equal)
```

_ie:_ it worsened for size=[128,1024]. not sure why.
but I _expect_ those sizes to be the more common for HEP workloads.
so I'd be enclined to leave the perf-improvements for sizes=[2,4,8,1024*1024] on the table.

what do you think @rmadar?