@rmadar 2020-04-24
ok let me re-phrase to make sure I understand:
 + option (1): we first fill `dst[0]` with `src[indices[0]]` and then do the loop starting from 1
 + option (2): we first do the loop up to `n` and then fill `dst[n]` with `src[indices[n]]` 

with:
 + option (1) is faster by ~ 100% wrt the very first implementation
 + option (2) is faster by ~15-20% wrt option (1) only for n = [128, 1024], but 30% worth for higher n (and actually similarish for n<128).

In that case, I would agree with you. Alternatively we could also use both calculation depending on n (with two internal functions which can be called in the proper case, ie n<1028 or n>1028 - I neglect the few % improvement for smaller sizes). But I don't know if this a "clean" option for an official software.