@sbinet 2020-04-24
> [rephrasing]

yes. you got it.

and picking various algorithm strategies depending on the inputs is what is generally done in various libraries, but one does really want the benefits to outweight the added "sophistication".

I'll leave it like is (_ie:_ sticking with the "i=1->n-1" iteration") and adding a new issue linking to this PR.